--- a/deps.ts
+++ b/deps.ts
@@ -4,4 +4,3 @@ export { assertEquals } from "https://deno.land/std@0.99.0/testing/asserts.ts";
 export * as esbuild from "https://deno.land/x/esbuild@v0.11.10/mod.js";
 export { AssertionError } from "https://deno.land/std@0.99.0/testing/asserts.ts";
 export { bench, runBenchmarks } from "https://deno.land/std@0.99.0/testing/bench.ts";
-export { createRequire } from "https://deno.land/std@0.99.0/node/module.ts";