… scripts.yml
… scripts/sk_bench.ts
… tools/gen_parser/skulpt_parser_genarator.py