close #77 


Bench marks after 50 runs of parsing turtle.py

| _ | master | this branch | skulpt | rust python |
|-|-|-|-|-|
| initial |  249 ms | 130ms | 131ms | 195ms
| final   | 151 ms | 48ms | 60ms | 195ms
| avg | 160 ms | 52ms | 64ms | 195ms

---
- `expect` checks the `type` matches by number rather than by string
- inline `mark()` and `reset()`
- `Tokenizer` is only responsible for getting the next token and making the `parser` responsible for `mark` meant it also took responsibility for `peek`, `getnext` and `diagnose`
- don't memoize parser methods like `expect`, `string` etc...
---

Which means we're now faster than skulpt 🎉 🐎